refactor: SessionStateExt and SessionConfigExt
moving them out from util to their own module as
util is too busy, also it will align with naming
in used in client.

also, split `SessionConfigExt` to `SessionConfigExt`
and `SessionConfigHelperExt`. `SessionConfigExt`
will stay user facing, `SessionConfigHelperExt`
will be used internally. This reduces ballista
client public api
