+++ ballista/core/src/lib.rs
@@ -32,6 +32,7 @@ pub mod consistent_hash;
 pub mod error;
 pub mod event_loop;
 pub mod execution_plans;
+pub mod extension;
 pub mod registry;
 pub mod utils;
